Use phase as double in Doppler correction
This avoids quantization in frequency at high sample rates. With
a float, the effective quantization in frequency would be on the
order of 1e-7 parts of the sampling rate.

+++ lib/doppler_correction_impl.cc
@@ -137,7 +137,7 @@ int doppler_correction_impl::work(int noutput_items,
         }
         d_phase += freq;
         phase_wrap();
-        const gr_complex nco = gr_expj(-d_phase);
+        const gr_complex nco = gr_expj(-static_cast<float>(d_phase));
         // gr::fast_cc_multiply is not available in GNU Radio 3.8,
         // so we use the usual product.
         out[j] = in[j] * nco;
